pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-privacy-cam"
version = "0.1.0"
description = "Virtual privacy camera with face masking and background protection"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "opencv-python>=4.9.0",
  "mediapipe==0.10.9",
  "pyvirtualcam>=0.12.0",
  "customtkinter>=5.2.2",
  "Pillow>=10.3.0",
  "numpy>=1.26.4",
]

[project.scripts]
smart-privacy-cam = "smart_privacy_cam.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
